geostore service added
